Rewrite with type-classes
