--- core/src/main/scala/gnieh/diffson/DiffsonInstance.scala
@@ -1,28 +1,0 @@
-/*
-* This file is part of the diffson project.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-package gnieh.diffson
-
-abstract class DiffsonInstance[JsValue]
-  extends JsonDiffSupport[JsValue]
-  with JsonMergePatchSupport[JsValue]
-  with JsonPatchSupport[JsValue]
-  with JsonPointerSupport[JsValue]
-  with JsonSupport[JsValue]
-  with LowPriorityImplicits[JsValue] {
-
-  val provider: JsonProvider
-
-}
