+++ core/src/test/scala/diffson/TestDynLcs.scala
@@ -1,0 +1,12 @@
+package diffson.lcs
+package test
+
+import cats.implicits._
+
+import org.scalatest._
+
+class TestDynLcs extends TestLcs {
+
+  val lcsImpl = new DynamicProgLcs[Char]
+
+}
