+++ core/src/test/scala/diffson/TestJsonMergePatch.scala
@@ -1,0 +1,41 @@
+package diffson
+package jsonmergepatch
+package test
+
+import cats.implicits._
+
+import org.scalatest._
+
+import scala.util.Try
+
+abstract class TestJsonMergePatch[Json](implicit Json: Jsony[Json]) extends FlatSpec with Matchers {
+
+  def parseJson(s: String): Json
+  def parsePatch(s: String): JsonMergePatch[Json]
+
+  val samples = List(
+    ("""{"a":"b"}""", """{"a":"c"}""", """{"a":"c"}"""),
+    ("""{"a":"b"}""", """{"b":"c"}""", """{"a":"b","b":"c"}"""),
+    ("""{"a":"b"}""", """{"a":null}""", """{}"""),
+    ("""{"a":"b","b":"c"}""", """{"a":null}""", """{"b":"c"}"""),
+    ("""{"a":["b"]}""", """{"a":"c"}""", """{"a":"c"}"""),
+    ("""{"a":"c"}""", """{"a":["b"]}""", """{"a":["b"]}"""),
+    ("""{"a": {"b": "c"}}""", """{"a": {"b": "d","c": null}}""", """{"a": {"b": "d"}}"""),
+    ("""{"a": [{"b":"c"}]}""", """{"a": [1]}""", """{"a": [1]}"""),
+    ("""["a","b"]""", """["c","d"]""", """["c","d"]"""),
+    ("""{"a":"b"}""", """["c"]""", """["c"]"""),
+    ("""{"a":"foo"}""", """null""", """null"""),
+    ("""{"a":"foo"}""", """"bar"""", """"bar""""),
+    ("""{"e":null}""", """{"a":1}""", """{"e":null,"a":1}"""),
+    ("""[1,2]""", """{"a":"b","c":null}""", """{"a":"b"}"""),
+    ("""{}""", """{"a":{"bb":{"ccc": null}}}""", """{"a":{"bb":{}}}"""))
+
+  for ((original, patch, result) <- samples)
+    s"patching $original with $patch" should s"result in $result" in {
+      val orig = parseJson(original)
+      val p = parsePatch(patch)
+      val res = parseJson(result)
+      p[Try](orig).get should be(res)
+    }
+
+}
