+++ core/src/test/scala/diffson/TestPatience.scala
@@ -1,0 +1,12 @@
+package diffson.lcs
+package test
+
+import cats.implicits._
+
+import org.scalatest._
+
+class TestPatience extends TestLcs {
+
+  val lcsImpl = new Patience[Char]
+
+}
