--- core/src/test/scala/gnieh/diffson/TestDynLcs.scala
@@ -1,10 +1,0 @@
-package gnieh.diffson
-package test
-
-import org.scalatest._
-
-class TestDynLcs extends TestLcs {
-
-  val lcsImpl = new DynamicProgLcs[Char]
-
-}
