--- core/src/test/scala/gnieh/diffson/TestPatience.scala
@@ -1,10 +1,0 @@
-package gnieh.diffson
-package test
-
-import org.scalatest._
-
-class TestPatience extends TestLcs {
-
-  val lcsImpl = new Patience[Char]
-
-}
