+++ project/plugins.sbt
@@ -1,9 +1,9 @@
-addSbtPlugin("com.typesafe.sbt" % "sbt-osgi" % "0.9.3")
+addSbtPlugin("com.typesafe.sbt" % "sbt-osgi" % "0.9.4")
 
 addSbtPlugin("org.scalariform" % "sbt-scalariform" % "1.8.2")
 
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.1")
 
-addSbtPlugin("com.eed3si9n" % "sbt-unidoc" % "0.4.1")
+addSbtPlugin("com.eed3si9n" % "sbt-unidoc" % "0.4.2")
 
 addSbtPlugin("com.codacy" % "sbt-codacy-coverage" % "1.3.15")
